+++ include/common.h
@@ -14,8 +14,9 @@ enum Mode {
 };
 
 
+void merge(float result[], const float left[], const float right[], const int left_size, const int right_size);
 void merge(float arr[], const int left_size, const int right_size);
 void mergeSort(float arr[], const int size);
 
-void run_original(const float data[], const int len, float& sum_value, float& max_value, float result[]);
-void run_speedup(const float data[], const int len, float& sum_value, float& max_value, float result[]);
+void run_original(const float data[], const int len, float& sum_value, float& max_value, float result[], double& sum_time, double& max_time, double &sort_time);
+void run_speedup(const float data[], const int len, float& sum_value, float& max_value, float result[], double& sum_time, double& max_time, double &sort_time);